Relocate Bezier tools from separate module to t2_charstrings.py
Bezier path related code previously in 'beziers_tools.py' has been moved into 't2_charstrings.py'. This refactor also includes renaming 'from_true_type' function to 'quadratics_to_cubics'.
