--- foundrytools_cli_2/lib/beziers/__init__.py
@@ -1,3 +1,0 @@
-from .beziers_tools import bezier_to_charstring
-
-__all__ = ["bezier_to_charstring"]
